+++ api/src/main/kotlin/com/retoday/api/domain/history/dto/response/HistoryRecordBatchResponse.kt
@@ -1,5 +1,6 @@
 package com.retoday.api.domain.history.dto.response
 
+import com.retoday.core.domain.history.dto.result.BatchResult
 import com.retoday.core.domain.history.dto.result.HistoryRecordBatchResult
 
 data class HistoryRecordBatchResponse(
